src/requirements.txt
fastapi
python-consul
python-json-logger
